Refatorando antes de comecar a Sprint 3 - V2
