+++ src/componentes/Despesas/CadastroDeDespesas/CadastroForm.js
@@ -1,10 +1,9 @@
 import React, {Component, Fragment, useContext, useEffect, useState} from "react";
 import {Formik, FieldArray, Field} from "formik";
-import { YupSignupSchemaCadastroDespesa, cpfMaskContitional, calculaValorRecursoAcoes, trataNumericos, convertToNumber, round, } from "../../../utils/ValidacoesAdicionaisFormularios";
+import { YupSignupSchemaCadastroDespesa, validaPayloadDespesas, validateFormDespesas, cpfMaskContitional, calculaValorRecursoAcoes,  } from "../../../utils/ValidacoesAdicionaisFormularios";
 import MaskedInput from 'react-text-mask'
 import { getDespesasTabelas, criarDespesa, alterarDespesa, deleteDespesa, getEspecificacoesCapital, getEspecificacoesCusteio} from "../../../services/Despesas.service";
 import {DatePickerField} from "../../DatePickerField";
-import moment from "moment";
 import {Button, Modal} from "react-bootstrap";
 import {useHistory} from 'react-router-dom'
 import {CadastroFormCusteio} from "./CadastroFormCusteio";
@@ -47,7 +46,7 @@ class DeletarModal extends Component {
             <Fragment>
                 <Modal centered show={this.props.show} onHide={this.handleClose}>
                     <Modal.Header closeButton>
-                        <Modal.Title>Deseja exluir está Despesa?</Modal.Title>
+                        <Modal.Title>Deseja excluir está Despesa?</Modal.Title>
                     </Modal.Header>
                     <Modal.Body>
                         <p>Tem certeza que deseja excluir esta despesa? A ação não poderá ser desfeita.</p>
@@ -57,7 +56,7 @@ class DeletarModal extends Component {
                             OK
                         </Button>
                         <Button variant="primary" onClick={this.props.handleClose}>
-                            fechar
+                            Fechar
                         </Button>
                     </Modal.Footer>
                 </Modal>
@@ -114,99 +113,7 @@ export const CadastroForm = () => {
 
     const onSubmit = async (values, {resetForm}) => {
 
-        // Quando é Alteração
-        if (typeof values.associacao === "object"){
-            values.associacao = localStorage.getItem(ASSOCIACAO_UUID)
-        }
-
-        if (typeof values.tipo_documento === "object" && values.tipo_documento !== null){
-            values.tipo_documento = values.tipo_documento.id
-        }else {
-            if (values.tipo_documento !== "" && values.tipo_documento !== "0" && values.tipo_documento !== 0 && values.tipo_documento !== null) {
-                values.tipo_documento = convertToNumber(values.tipo_documento);
-            } else {
-                values.tipo_documento = null
-            }
-        }
-
-        if (typeof values.tipo_transacao === "object" && values.tipo_transacao !== null){
-            values.tipo_transacao = values.tipo_transacao.id
-        }else {
-            if (values.tipo_transacao !== "" && values.tipo_transacao !== "0" && values.tipo_transacao !== 0 && values.tipo_transacao !== null) {
-                values.tipo_transacao = convertToNumber(values.tipo_transacao);
-            } else {
-                values.tipo_transacao = null
-            }
-        }
-
-        values.valor_total = trataNumericos(values.valor_total);
-        values.valor_recursos_proprios = trataNumericos(values.valor_recursos_proprios);
-        values.valor_recusos_acoes = round((values.valor_recusos_acoes), 2)
-
-        if (values.data_documento !== "" && values.data_documento !== null){
-            values.data_documento = moment(values.data_documento).format("YYYY-MM-DD");
-        }else {
-            values.data_documento = null
-        }
-
-        if (values.data_transacao !== "" && values.data_transacao !== null){
-            values.data_transacao = moment(values.data_transacao).format("YYYY-MM-DD");
-        }else {
-            values.data_transacao = null
-        }
-
-        values.rateios.map((rateio) => {
-
-            if (typeof rateio.especificacao_material_servico === "object" && rateio.especificacao_material_servico !== null){
-                rateio.especificacao_material_servico = rateio.especificacao_material_servico.id
-            }else {
-                rateio.especificacao_material_servico = convertToNumber(rateio.especificacao_material_servico)
-            }
-
-            if (typeof rateio.conta_associacao === "object" && rateio.conta_associacao !== null){
-                rateio.conta_associacao = rateio.conta_associacao.uuid
-            }else {
-                if (rateio.conta_associacao === "0" || rateio.conta_associacao === "" || rateio.conta_associacao === 0){
-                    rateio.conta_associacao = null
-                }
-            }
-
-            if (typeof rateio.acao_associacao === "object" && rateio.acao_associacao !== null){
-                rateio.acao_associacao = rateio.acao_associacao.uuid
-            }else {
-                if (rateio.acao_associacao === "0" || rateio.acao_associacao === "" || rateio.acao_associacao === 0) {
-                    rateio.acao_associacao = null
-                }
-            }
-
-            if (typeof rateio.tipo_custeio === "object" && rateio.tipo_custeio !== null){
-                rateio.tipo_custeio = rateio.tipo_custeio.id
-            }else {
-
-                if (rateio.tipo_custeio === "0" || rateio.tipo_custeio === 0 || rateio.tipo_custeio === ""){
-                    rateio.tipo_custeio = null
-                }else {
-                    rateio.tipo_custeio = convertToNumber(rateio.tipo_custeio)
-                }
-            }
-
-            rateio.quantidade_itens_capital = convertToNumber(rateio.quantidade_itens_capital)
-            rateio.valor_item_capital = trataNumericos(rateio.valor_item_capital)
-            rateio.valor_rateio = round(trataNumericos(rateio.valor_rateio),2)
-
-            if (rateio.aplicacao_recurso === "0" || rateio.aplicacao_recurso === "" || rateio.aplicacao_recurso === 0){
-                rateio.aplicacao_recurso = null
-            }
-
-            if (rateio.especificacao_material_servico === "0" || rateio.especificacao_material_servico === 0 || rateio.especificacao_material_servico === ""){
-                rateio.especificacao_material_servico = null
-            }
-
-            if (rateio.aplicacao_recurso === "CAPITAL"){
-                rateio.valor_rateio = round(rateio.quantidade_itens_capital * rateio.valor_item_capital, 2)
-            }
-
-        })
+        validaPayloadDespesas(values)
 
         console.log("onSubmit", values)
 
@@ -280,42 +187,6 @@ export const CadastroForm = () => {
         });
     }
 
-    // Synchronous validation
-    const validate = (values, props /* only available when using withFormik */) => {
-        const errors = {};
-
-        console.log("Entei no validade", values)
-
-        let var_valor_recursos_acoes = trataNumericos(values.valor_total) - trataNumericos(values.valor_recursos_proprios)
-        console.log(" Var Recurso Acoes ", var_valor_recursos_acoes)
-
-        let var_valor_total_dos_rateios = 0;
-        let var_valor_total_dos_rateios_capital = 0;
-        let var_valor_total_dos_rateios_custeio = 0;
-
-        values.rateios.map((rateio) => {
-            if (rateio.aplicacao_recurso === "CAPITAL"){
-                var_valor_total_dos_rateios_capital = var_valor_total_dos_rateios_capital + round(trataNumericos(rateio.quantidade_itens_capital) * trataNumericos(rateio.valor_item_capital, 2))
-            }else{
-                var_valor_total_dos_rateios_custeio = var_valor_total_dos_rateios_custeio + trataNumericos(rateio.valor_rateio)
-            }
-        })
-
-        console.log(" Var Valor Total dos Rateios CAPITAL ", var_valor_total_dos_rateios_capital)
-        console.log(" Var Valor Total dos Rateios CUSTEIO ", var_valor_total_dos_rateios_custeio)
-
-        var_valor_total_dos_rateios = var_valor_total_dos_rateios_capital + var_valor_total_dos_rateios_custeio
-
-        console.log(" Var Valor Total dos Rateios ", var_valor_total_dos_rateios)
-
-
-        if (var_valor_recursos_acoes !== var_valor_total_dos_rateios) {
-            errors.valor_recusos_acoes = 'O total das classificações deve corresponder ao valor total da nota';
-        }
-        return errors;
-    };
-
-
     return (
         <>
             <Formik
@@ -324,14 +195,13 @@ export const CadastroForm = () => {
                 validateOnBlur={false}
                 onSubmit={onSubmit}
                 enableReinitialize={true}
-                validate={validate}
-                //onReset={(props) => handleReset(props)}
+                validate={validateFormDespesas}
             >
                 {props => {
                     const {
                         values,
                         setFieldValue,
-                        errors, touched, isValidating
+                        errors,
                     } = props;
                     return (
                         <form onSubmit={props.handleSubmit}>
@@ -370,7 +240,6 @@ export const CadastroForm = () => {
                                                 props.values.tipo_documento === "object" ? props.values.tipo_documento.id : props.values.tipo_documento.id
                                             ) : 0
                                         }
-                                        //value={props.values.tipo_documento !== null ? props.values.tipo_documento.id : 0 }
                                         onChange={props.handleChange}
                                         onBlur={props.handleBlur}
                                         name='tipo_documento'
@@ -409,7 +278,6 @@ export const CadastroForm = () => {
                                 <div className="col-12 col-md-3 mt-4">
                                     <label htmlFor="tipo_transacao">Tipo de transação</label>
                                     <select
-                                        //value={props.values.tipo_transacao !== null ? props.values.tipo_transacao.id : 0}
                                         value={
                                             props.values.tipo_transacao !== null ? (
                                                 props.values.tipo_transacao === "object" ? props.values.tipo_transacao.id : props.values.tipo_transacao.id
@@ -436,7 +304,6 @@ export const CadastroForm = () => {
                                         name="data_transacao"
                                         id="data_transacao"
                                         value={values.data_transacao != null ? values.data_transacao : ""}
-                                        //value={values.data_transacao}
                                         onChange={setFieldValue}
                                     />
                                     {props.errors.data_transacao &&
